+++ nodejs-server/service/ArgoneService.js
@@ -1,5 +1,8 @@
 'use strict';
-
+const argone = require('../models/argone');
+const GJV = require('geojson-validation');
+const pointInPolygon = require('@turf/boolean-point-in-polygon').default;
+const helpers = require('../helpers/helpers')
 
 /**
  * Probabilities of floats moving between two points in a range of forecast projections
@@ -12,34 +15,78 @@
  * data List Forecast durations to include. Return only documents that have all data requested. Accepts ~ negation to filter out documents including the specified data. Omission of this parameter will result in metadata only responses. (optional)
  * returns List
  **/
-exports.findargone = function(id,forecastOrigin,forecastGeolocation,metadata,compression,data) {
+exports.findargone = function(res, id,forecastOrigin,forecastGeolocation,metadata,compression,data) {
   return new Promise(function(resolve, reject) {
-    var examples = {};
-    examples['application/json'] = [ {
-  "metadata" : [ "metadata", "metadata" ],
-  "data" : [ [ "", "" ], [ "", "" ] ],
-  "_id" : "_id",
-  "geolocation" : {
-    "coordinates" : [ 0.8008281904610115, 0.8008281904610115 ],
-    "type" : "type"
-  }
-}, {
-  "metadata" : [ "metadata", "metadata" ],
-  "data" : [ [ "", "" ], [ "", "" ] ],
-  "_id" : "_id",
-  "geolocation" : {
-    "coordinates" : [ 0.8008281904610115, 0.8008281904610115 ],
-    "type" : "type"
-  }
-} ];
-    if (Object.keys(examples).length > 0) {
-      resolve(examples[Object.keys(examples)[0]]);
-    } else {
-      resolve();
+
+    // decide y/n whether to service this request; sanitize inputs
+    if(!forecastOrigin && !forecastGeolocation && !id){
+        reject({"code": 400, "message": "please specify at least one of forecastOrigin, forecastGeolocation and/or id"})
+        return
     }
+    if(forecastOrigin){
+        forecastOrigin = {"type": "Point", "coordinates": forecastOrigin}
+        if (!GJV.valid(forecastOrigin) || !GJV.isPoint(forecastOrigin)){
+            reject({"code": 400, "message": "forecastOrigin must be a valid <longitude,latitude>"});
+            return;
+        }
+    }
+    if(forecastGeolocation){
+        forecastGeolocation = {"type": "Point", "coordinates": forecastGeolocation}
+        if (!GJV.valid(forecastGeolocation) || !GJV.isPoint(forecastGeolocation)){
+            reject({"code": 400, "message": "forecastGeolocation must be a valid <longitude,latitude>"});
+            return;
+        }
+    }
+
+    // local filter: fields in data collection other than geolocation and timestamp 
+    let local_filter = []
+    if(forecastOrigin){
+        local_filter.push({'$geoNear': {'near': forecastOrigin, 'maxDistance': 1, 'distanceField': 'dist',  'key': 'geolocation'}})
+    }
+    if(forecastGeolocation){
+        local_filter.push({'$geoNear': {'near': forecastGeolocation, 'maxDistance': 1, 'distanceField': 'dist',  'key': 'geolocation_forecast'}})
+    }
+    if(id){
+        local_filter.push({'$match':{'_id': id}})
+    }
+
+    // postprocessing parameters
+    let pp_params = {
+        compression: compression,
+        data: JSON.stringify(data) === '["except-data-values"]' ? null : data, // ie `data=except-data-values` is the same as just omitting the data qsp
+        junk: ['dist']
+    }
+
+    // metadata table filter: no-op promise stub, nothing to filter grid data docs on from metadata at the moment
+    let metafilter = Promise.resolve([])
+    let metacomplete = false
+
+    // datafilter must run syncronously after metafilter in case metadata info is the only search parameter for the data collection
+    let datafilter = metafilter.then(helpers.datatable_stream.bind(null, argone['argone'], {}, local_filter))
+
+    Promise.all([metafilter, datafilter])
+        .then(search_result => {
+          let stub = function(data, metadata){
+              // given a data and corresponding metadata document,
+              // return the record that should be returned when the compression=minimal API flag is set
+              // should be id, long, lat, timestamp, and then anything needed to group this point together with other points in interesting ways.
+              return [
+                data['_id'],
+                data.geolocation.coordinates[0], 
+                data.geolocation.coordinates[1], 
+                data.geolocation_forecast.coordinates[0], 
+                data.geolocation_forecast.coordinates[1]
+              ]
+          }
+
+          let postprocess = helpers.post_xform(argone['argoneMeta'], pp_params, search_result, res, stub)
+
+          resolve([search_result[1], postprocess])
+
+        })
+
   });
 }
-
 
 /**
  * argone metadata search and filter.
@@ -47,49 +94,19 @@ exports.findargone = function(id,forecastOrigin,forecastGeolocation,metadata,com
  * id String Unique ID to search for. (optional)
  * returns List
  **/
-exports.findargoneMeta = function(id) {
+exports.findargoneMeta = function(res,id) {
   return new Promise(function(resolve, reject) {
-    var examples = {};
-    examples['application/json'] = [ {
-  "data_info" : [ "", "" ],
-  "data_type" : "data_type",
-  "_id" : "_id",
-  "source" : [ {
-    "date_updated" : "2000-01-23T04:56:07.000+00:00",
-    "source" : [ "source", "source" ],
-    "url" : "url",
-    "doi" : "doi"
-  }, {
-    "date_updated" : "2000-01-23T04:56:07.000+00:00",
-    "source" : [ "source", "source" ],
-    "url" : "url",
-    "doi" : "doi"
-  } ],
-  "date_updated_argovis" : "2000-01-23T04:56:07.000+00:00",
-  "levels" : [ 0.8008281904610115, 0.8008281904610115 ]
-}, {
-  "data_info" : [ "", "" ],
-  "data_type" : "data_type",
-  "_id" : "_id",
-  "source" : [ {
-    "date_updated" : "2000-01-23T04:56:07.000+00:00",
-    "source" : [ "source", "source" ],
-    "url" : "url",
-    "doi" : "doi"
-  }, {
-    "date_updated" : "2000-01-23T04:56:07.000+00:00",
-    "source" : [ "source", "source" ],
-    "url" : "url",
-    "doi" : "doi"
-  } ],
-  "date_updated_argovis" : "2000-01-23T04:56:07.000+00:00",
-  "levels" : [ 0.8008281904610115, 0.8008281904610115 ]
-} ];
-    if (Object.keys(examples).length > 0) {
-      resolve(examples[Object.keys(examples)[0]]);
-    } else {
-      resolve();
+    if(id !== 'covariance'){
+      reject({
+        code: 404,
+        message: "No float location metadata matching ID " + id + "; all float location metadata is stored in the single document id=covariance"
+      })
     }
+    const query = argone['argoneMeta'].aggregate([{$match:{'_id':id}}]);
+    let postprocess = helpers.meta_xform(res)
+    resolve([query.cursor(), postprocess])
   });
 }
+
+
